+++ client/poetry/src/components/Header.js
@@ -1,4 +1,5 @@
 
+import { Link } from 'react-router-dom'
 
 function Header() {
     return (
@@ -6,7 +7,9 @@ function Header() {
        
   <div className="logo">شعر</div>
   <div className="title">תרגום שירה ערבית</div>
+  <Link to="/">חזרה לראשי</Link>
       </div>
+      
     );
   }
