+++ client/poetry/src/components/Songlist.js
@@ -15,27 +15,27 @@ const Songlist = () => {
       //   "https://6069b5a1e1c2a10017544d37.mockapi.io/products"
       // );
 
-      // const data = [{title:"שירי אהבה"},{title:"שירי געגוע"},{title:"שירים קצרים"}]
-      const data =  [
-        'قصائد اعتذار',   'قصائد الاناشيد',
-        'قصائد المعلقات', 'قصائد حزينه',
-        'قصائد دينية',    'قصائد ذم',
-        'قصائد رثاء',     'قصائد رومنسيه',
-        'قصائد سياسية',   'قصائد شوق',
-        'قصائد عامه',     'قصائد عتاب',
-        'قصائد غزل',      'قصائد فراق',
-        'قصائد قصيره',    'قصائد مدح',
-        'قصائد هجاء',     'قصائد وطنيه'
-      ]
+      const data = [{title:"שיר 1"},{title:"שיר 2"},{title:"שיר 3"}]
+      // const data =  [
+      //   'قصائد اعتذار',   'قصائد الاناشيد',
+      //   'قصائد المعلقات', 'قصائد حزينه',
+      //   'قصائد دينية',    'قصائد ذم',
+      //   'قصائد رثاء',     'قصائد رومنسيه',
+      //   'قصائد سياسية',   'قصائد شوق',
+      //   'قصائد عامه',     'قصائد عتاب',
+      //   'قصائد غزل',      'قصائد فراق',
+      //   'قصائد قصيره',    'قصائد مدح',
+      //   'قصائد هجاء',     'قصائد وطنيه'
+      // ]
  
       setTerm(data.map(topic => {
         
         return (    
     
-      <Link to={{pathname: `topics/${topic.title}`, query: topic}}
+      <Link to={{pathname: `/Songpage`, query: topic}}
       className="links" key={topic.title}>
         
-        {topic}<br/>
+        {topic.title}<br/>
         </Link>
 
           )
